Move result_unit_err to its own module
